requirements.txt
Flask>=3.0,<4.0
pandas>=2.2,<3.0
matplotlib>=3.8,<4.0
Pillow>=10.0,<12.0
numpy>=1.26,<3.0
pydub>=0.25,<1.0
scikit-learn>=1.4,<2.0
